Set it up to try and take a reading every minute, and can reconnect if the connection to the mifi goes down. Also prompts user for background use
